requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
googleapis-common-protos==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.1
langfuse==3.8.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
